fix: add unit test
